pyproject.toml
[project]
name = "seda"
version = "0.1.0"
description = "State-linked Entity Detection & Analysis - Iranian regime-affiliated Twitter account detection"
requires-python = ">=3.11"
dependencies = [
    "apify-client>=1.6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "lightgbm>=4.0",
    "hazm>=0.9",
    "transformers>=4.35",
    "torch>=2.0",
    "networkx>=3.0",
    "python-louvain>=0.16",
    "datasketch>=1.6",
    "anthropic>=0.18",
    "streamlit>=1.30",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "plotly>=5.18",
    "typer>=0.9",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "libsql-experimental>=0.0.47",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.scripts]
seda-collect = "scripts.collect:app"
seda-analyze = "scripts.analyze:app"
seda-seeds = "scripts.seed_accounts:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["seda*", "scripts*", "dashboard*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
